refactor(workspace): promote atomic write helper to shared utils::fs
Move the #2318 workspace atomic writer (temp-file + sync_all + rename) out of
workspace/atomic.rs into utils::fs::write_atomic so every config-persistence
path can share one implementation, and point the workspace callers at it.
No behavior change; the helper's tests move with it.
