+++ src-tauri/src/workspace/mod.rs
@@ -1,4 +1,3 @@
-mod atomic;
 pub mod config;
 pub mod last_session;
 pub mod manager;
